requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.10
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
